API BREAK: xdot: make 'xdot_linear_grad.n_stops' unsigned
This pays down some debt incurred in
50c35df8b3943749d529f9e97d768ef5f9b71fb4. It introduces some cast noise,
but this will be removed in a follow-on commit.

Note that this change is still backwards compatible, in that `n_stops`
is expected to always be ≥0, so libxdot after this commit will still
be able to parse valid xdot text produced by libxdot before this commit.
Similarly xdot text produced by libxdot after this commit will still be
parseable by libxdot before this commit, provided the number of stops
does not exceed `INT_MAX`.

+++ plugin/core/gvrender_core_json.c
@@ -192,7 +192,7 @@ static void write_linear_grad (GVJ_t * job, xdot_linear_grad* lg, state_t* sp)
     indent(job, sp->Level);
     gvprintf(job, "\"p1\": [%.03f,%.03f],\n", lg->x1, lg->y1); 
     indent(job, sp->Level);
-    write_stops (job, lg->n_stops, lg->stops, sp);
+    write_stops(job, (int)lg->n_stops, lg->stops, sp);
 }
 
 static void write_xdot (xdot_op * op, GVJ_t * job, state_t* sp)
